:white_check_mark: Move unit tests into own unit subdirectory
