+++ package.json
@@ -146,15 +146,15 @@
     "build": "yarn clean && rm -rf dist && tsc && yarn build-client",
     "build-client": "webpack",
     "clean": "rm -f .oclif.manifest.json",
-    "lint": "tsc -p test --noEmit && tslint -p test -t stylish --fix",
+    "lint": "tsc -p test/unit --noEmit && tslint -p test/unit -t stylish --fix",
     "postpublish": "yarn clean",
     "posttest": "yarn lint",
     "prepublishOnly": "yarn build && oclif-dev manifest && oclif-dev readme",
     "preversion": "yarn clean",
-    "test": "yarn test:unit && yarn test:client && yarn test:integration && yarn test:command:snapshot && yarn test:command:upload",
-    "test:unit": "PERCY_TOKEN=abc mocha \"test/**/*.test.ts\" --exclude \"test/percy-agent-client/**/*.test.ts\" --exclude \"test/integration/**/*\"",
-    "test:client": "karma start ./test/percy-agent-client/karma.conf.js",
-    "test:integration": "yarn build-client && node ./bin/run exec --cache-responses -h *.localtest.me -c .ci.percy.yml -- mocha test/integration/**/*.test.ts",
+    "test": "yarn test:unit && yarn test:client && yarn test:acceptance && yarn test:integration && yarn test:command:snapshot && yarn test:command:upload",
+    "test:unit": "PERCY_TOKEN=abc mocha \"test/**/*.test.ts\" --exclude \"test/unit/percy-agent-client/**/*.test.ts\" --exclude \"test/integration/**/*\" --config test/unit/.mocha.yml",
+    "test:client": "karma start ./test/unit/percy-agent-client/karma.conf.js",
+    "test:integration": "yarn build-client && node ./bin/run exec --cache-responses -h *.localtest.me -c .ci.percy.yml -- mocha test/integration/**/*.test.ts --config test/unit/.mocha.yml",
     "test:command:snapshot": "./bin/run snapshot test/integration/test-static-site -b /dummy-base-url/ -i **/red-keep.** -s **/*.{html}",
     "test:command:upload": "./bin/run upload test/integration/test-static-images",
     "version": "oclif-dev readme && git add README.md",
