+++ test/acceptance/.mocha.yml
@@ -1,0 +1,4 @@
+require:
+  - "@babel/register"
+  - "regenerator-runtime/runtime"
+timeout: 5000
