+++ test/acceptance/dummy/snapshot.js
