+++ test/acceptance/helpers/snapshot.js
