+++ test/unit/commands/finalize.test.ts
