+++ test/unit/commands/snapshot.test.ts
@@ -1,10 +1,10 @@
 import { expect, test } from '@oclif/test'
 import { describe } from 'mocha'
 import * as sinon from 'sinon'
-import Snapshot from '../../src/commands/snapshot'
-import { DEFAULT_CONFIGURATION } from '../../src/configuration/configuration'
-import { AgentService } from '../../src/services/agent-service'
-import StaticSnapshotService from '../../src/services/static-snapshot-service'
+import Snapshot from '../../../src/commands/snapshot'
+import { DEFAULT_CONFIGURATION } from '../../../src/configuration/configuration'
+import { AgentService } from '../../../src/services/agent-service'
+import StaticSnapshotService from '../../../src/services/static-snapshot-service'
 import { captureStdErr, captureStdOut } from '../helpers/stdout'
 import chai from '../support/chai'
