+++ test/unit/fixtures/services/static-snapshot-service/_dummy-testing-app/index.js
