+++ test/unit/helpers/init.ts
@@ -1,3 +1,3 @@
 // tslint:disable-next-line:no-var-requires
 const path = require('path')
-process.env.TS_NODE_PROJECT = path.resolve('test/tsconfig.json')
+process.env.TS_NODE_PROJECT = path.resolve('test/unit/tsconfig.json')
