+++ test/unit/services/agent-service.test.ts
@@ -1,7 +1,7 @@
 import {describe} from 'mocha'
 import * as nock from 'nock'
-import { DEFAULT_CONFIGURATION } from '../../src/configuration/configuration'
-import {AgentService} from '../../src/services/agent-service'
+import { DEFAULT_CONFIGURATION } from '../../../src/configuration/configuration'
+import {AgentService} from '../../../src/services/agent-service'
 import {captureStdOut} from '../helpers/stdout'
 import chai from '../support/chai'
 const expect = chai.expect
