+++ test/unit/services/process-service.test.ts
@@ -1,6 +1,6 @@
 import {expect} from 'chai'
 import {describe} from 'mocha'
-import ProcessService from '../../src/services/process-service'
+import ProcessService from '../../../src/services/process-service'
 import {createPidFile, deletePidFile} from '../helpers/process'
 
 describe('ProcessService', () => {
