+++ test/unit/tsconfig.json
@@ -1,10 +1,11 @@
 {
-  "extends": "../tsconfig",
+  "extends": "../../tsconfig",
   "compilerOptions": {
     "sourceMap": true
   },
   "include": [
     "./**/*",
-    "../src/**/*"
+    "../integration/**/*",
+    "../../src/**/*"
   ]
 }
